Add custom object hooks
This adds a functionality for overriding/customising the structured object
logged at request received/success/error. It is *like* the
custom[Received|Success|Error]Message hooks, but works on customising the
structured object.

It was opted not to overload *Message due to that being specifically about
messages, and the risk that adjusting that could break an existing usage
of that hook.

The README has been updated with example usage.