feat(ts): loosening up the IncomingMessage and ServerResponse typings
The typings on the various instances of `IncomingMessage` and `ServerResponse`
are too strict when coupled with Express and Fastify `Request` instances that
have been extended for custom properties. This loosens them up by allowing the
main `pino` instance to be supplied a generic of the Request and Response
interfaces that you use in your application.

Fixes #294